Updated test so it runs on py.test > 0.9.2

+++ test/test_zippack.py
@@ -6,42 +6,59 @@ import os
 
 import py.test
 
-from common import SamplePart
 from openpack.basepack import *
 from openpack.zippack import ZipPackage
+
+from common import SamplePart
 
 TESTFILE = os.path.join(os.path.abspath(os.path.dirname(__file__)), 'temp.zipx')
 
 class TestZipPack(object):
+	def setup_class(cls):
+		cls.remove_testfile()
+
 	def test_create(self):
 		self.pack = ZipPackage(TESTFILE)
 	
 	def test_add_part(self):
+		self.test_create()
 		self.part = p = SamplePart(self.pack, '/test/part.xml')
 		self.pack[p.name] = p
 		self.pack.content_types.add_override(p)
 		self.pack.relate(p)
 	
 	def test_write_to_part(self):
+		self.test_add_part()
 		self.part.data = '<test>hi there</test>'
 	
 	def test_save(self):
+		self.test_write_to_part()
 		self.pack.save()
 		del self.pack
 		del self.part
 	
 	def test_open(self):
+		self.test_save()
 		self.pack = ZipPackage(TESTFILE)
 	
 	def test_contents(self):
+		self.test_open()
 		assert '/test/part.xml' in self.pack
 		self.part = self.pack['/test/part.xml']
 		assert self.part.data == '<test>hi there</test>'
 	
 	def test_relationships(self):
+		self.test_contents()
 		print self.pack.relationships.children
 		assert self.pack.related('http://polimetrix.com/relationships/test')[0] == self.part
 
 	def teardown_class(cls):
-		os.remove(TESTFILE)
+		if hasattr(self, 'pack'): del self.pack
+		if hasattr(self, 'part'): del self.part
+		cls.remove_testfile()
+
+	@classmethod
+	def remove_testfile(cls):
+		if os.path.exists(TESTFILE):
+			os.remove(TESTFILE)
